refactor: move funtions into hook
